feat: ✨ implement additional tag functionality (#19)
* feat: :sparkles: add frontmatter tag field

add field `frontmatter_tags` which are formatted as quoted tags, useful to be put into frontmatter ```---\ntags: [ {{ frontmatter_tags }} ]\n```

* feat: :sparkles: add frontmatter tag field

add field `frontmatter_tags` which are formatted as quoted tags, useful to be put into frontmatter ```---\ntags: [ {{ frontmatter_tags }} ]\n```

* feat: :sparkles: roll-up of highlight tags into frontmatter

a rolled-up (deduplicated) list highlight tags can be used in frontmatter via the field `quoted_highlight_tags`

* refactor: :recycle: change field names to reflect their use in frontmatter

frontmatter tags should not be quoted, and tags can not contain whitespace. This change sanitizes tags and provides an option to return the tags without a hash (#).

* feat: :sparkles: user setting for "colon" (:) replacement (#22)

* feat: :sparkles: user setting for "colon" (:) replacement

The colon in a title is a special character which by default is replaced witha a dash (-) in filenames. With this change, users can choose what string to use as a replacement.

* feat: implement readwise throttling
due to cross-origin / request limitations, access to the "Retry-After" header is limited. However, readwise returns the throttling info in the body. This change implements the necessary parsing and catch if parsing wouldn't work.

* feat: :sparkles: add frontmatter tag field

add field `frontmatter_tags` which are formatted as quoted tags, useful to be put into frontmatter ```---\ntags: [ {{ frontmatter_tags }} ]\n```

* refactor: ⚗️ introduce option for quotes in tag format

- make tag formatting more flexible
- avoid adding quotes all the time
- differentiate nohas from quoting

* refactor: ⚗️ introduce option for quotes in tag format