feat: :sparkles: differentiate title from sanitized title
while the sanitized title is necessary for file names, one migth actually want to use the *real* title in the contents. This change implements this.

The old default header template uses `title` which will create invalid links for titles with special characters. Users are advised to update their templates before running a sync.