feat: :sparkles: implement new export api with cache
with this change, and using the `export` API, the plugin will keep a local cache of the full library and will integrate updated highlights into the cache. This also opens the door to 3-way merge operations as we will have the original, the loacl, and the remote Readwise changes available which should allow us to work towards more robust forms of merging.

major arcitectural change with respect to keeping a local cache.