feat: ✨ separate tracking from file renaming
add a new setting to enable file renames separate from tracking notes with a property. This enhances additional use cases, e.g. keeping stable filenames oin your Vault even if the title or some other metadata with relevance to the filename changes during a sync.

based on a discussion in #65.

CLOSES #65