refactor: 🏗️ add a composite pass to the nunjucks extension
when we don't atomize highlights (e.g. non-atomic notes when using `rw-atomize` for individual files), the frontmatter must be removed from the rendered template, otherwise it will show up in the Readwise note. We introduce a `COMPOSITE` pass which only returns a highlight's rendered content, but not the `frontmatter` block.